Moved used util functions to separate files
