+++ packages/components/src/Form/Inputs/InputColor/LightSaturationPreview/LightSaturationPreview.tsx
@@ -29,7 +29,7 @@ import React, { FC, MouseEvent, useEffect, useRef } from 'react'
 import styled from 'styled-components'
 import { useMouseDragPosition, usePreviousValue } from '../../../../utils'
 import { COLOR_PICKER_WIDTH } from '../ColorPicker/dimensions'
-import { simpleHSVtoFormattedColorString } from '../utils/color_format_utils'
+import { simpleHSVtoFormattedColorString } from '../utils'
 import { SimpleHSV } from '../types'
 
 const PREVIEW_HEIGHT = 150
