+++ packages/components/src/Form/Inputs/InputColor/utils/index.ts
@@ -24,19 +24,7 @@
 
  */
 
-import { toPercent } from './toPercent'
-
-describe('math_utils', () => {
-  const correctValues = [60, 24, 10, 13, 25]
-  test('toPercent', () => {
-    ;[
-      [60, 100],
-      [60, 255],
-      [25, 255],
-      [32, 255],
-      [96, 384],
-    ].map((values, index) =>
-      expect(toPercent(values[0], values[1])).toBe(correctValues[index])
-    )
-  })
-})
+export * from './color_utils'
+export * from './hsvToHex'
+export * from './simpleHsvToFormattedColorString'
+export * from './stringToSimpleHsv'
