+++ packages/components/src/Form/Inputs/InputColor/utils/rgbToColorName.ts
@@ -24,19 +24,17 @@
 
  */
 
-import { toPercent } from './toPercent'
+import { RGBColor, HSLColor, rgb } from 'd3-color'
+import { getOpacity } from './getOpacity'
+import { namedColorLookup } from './namedColorLookup'
 
-describe('math_utils', () => {
-  const correctValues = [60, 24, 10, 13, 25]
-  test('toPercent', () => {
-    ;[
-      [60, 100],
-      [60, 255],
-      [25, 255],
-      [32, 255],
-      [96, 384],
-    ].map((values, index) =>
-      expect(toPercent(values[0], values[1])).toBe(correctValues[index])
-    )
-  })
-})
+export const rgbToColorName = (
+  color: RGBColor | HSLColor,
+  opacity: number | null = null
+) => {
+  const opacityUse = opacity || getOpacity(color)
+  const name = namedColorLookup(color)
+  if (name) return name
+  if (opacityUse !== 1) return rgb(color).toString()
+  return color.hex()
+}
