+++ packages/components/src/Form/Inputs/InputColor/utils/toPercent.ts
